Basic random sandwich page.
