+++ api/src/controllers/sandwichController.js
@@ -1,0 +1,12 @@
+const fs = require('fs');
+
+// To view
+const randomView = (req, res) => {
+    res.render('sandwiches', {});
+};
+
+// Gallery View
+
+module.exports = {
+    randomView
+};
