+++ api/src/recs.json
@@ -1,127 +1,152 @@
 [
-    {
-        "id": 1,
-        "name": "Tuna Salad",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 2,
-        "name": "Tuna Salad",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 3,
-        "name": "Tuna Salad",
-        "restaurant": "Otherside Deli"
-    },
-    {
-        "id": 4,
-        "name": "Banh Mi",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 5,
-        "name": "Classic Pastrami",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 6,
-        "name": "Classic Pastrami",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 7,
-        "name": "Tuna Salad",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 8,
-        "name": "Harry's Perfect Pastrami",
-        "restaurant": "The Silver Medal"
-    },
-    {
-        "id": 9,
-        "name": "Banh Mi",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 10,
-        "name": "Banh Mi",
-        "restaurant": "The Silver Medal"
-    },
-    {
-        "id": 11,
-        "name": "Harry's Perfect Pastrami",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 12,
-        "name": "Banh Mi",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 13,
-        "name": "Fried Bologna",
-        "restaurant": "The Silver Medal"
-    },
-    {
-        "id": 14,
-        "name": "Tuna Salad",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 15,
-        "name": "Classic Pastrami",
-        "restaurant": "Otherside Deli"
-    },
-    {
-        "id": 16,
-        "name": "Fried Bologna",
-        "restaurant": "The Silver Medal"
-    },
-    {
-        "id": 17,
-        "name": "BLT",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 18,
-        "name": "BLT",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 19,
-        "name": "Tuna Salad",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 20,
-        "name": "Classic Pastrami",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 21,
-        "name": "Harry's Perfect Pastrami",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 22,
-        "name": "Classic Pastrami",
-        "restaurant": "Little Deli"
-    },
-    {
-        "id": 23,
-        "name": "Banh Mi",
-        "restaurant": "Pho Vans"
-    },
-    {
-        "id": 24,
-        "name": "Classic Pastrami",
-        "restaurant": "Otherside Deli"
-    },
-    {
-        "id": 25,
-        "name": "BLT",
-        "restaurant": "Little Deli"
-    }
+   {
+      "id": 1,
+      "name": "Tuna Salad",
+      "restaurant": "Pho Vans",
+      "blurb": "Best sandwich ever."
+   },
+   {
+      "id": 2,
+      "name": "Tuna Salad",
+      "restaurant": "Little Deli",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 3,
+      "name": "Tuna Salad",
+      "restaurant": "Otherside Deli",
+      "blurb": "It's fucking good."
+   },
+   {
+      "id": 4,
+      "name": "Banh Mi",
+      "restaurant": "Little Deli",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 5,
+      "name": "Classic Pastrami",
+      "restaurant": "Little Deli",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 6,
+      "name": "Classic Pastrami",
+      "restaurant": "Pho Vans",
+      "blurb": "It's fucking good."
+   },
+   {
+      "id": 7,
+      "name": "Tuna Salad",
+      "restaurant": "Pho Vans",
+      "blurb": "It's fucking good."
+   },
+   {
+      "id": 8,
+      "name": "Harry's Perfect Pastrami",
+      "restaurant": "The Silver Medal",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 9,
+      "name": "Banh Mi",
+      "restaurant": "Little Deli",
+      "blurb": "Best sandwich ever."
+   },
+   {
+      "id": 10,
+      "name": "Banh Mi",
+      "restaurant": "The Silver Medal",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 11,
+      "name": "Harry's Perfect Pastrami",
+      "restaurant": "Little Deli",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 12,
+      "name": "Banh Mi",
+      "restaurant": "Little Deli",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 13,
+      "name": "Fried Bologna",
+      "restaurant": "The Silver Medal",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 14,
+      "name": "Tuna Salad",
+      "restaurant": "Pho Vans",
+      "blurb": "yum yum eeatumup."
+   },
+   {
+      "id": 15,
+      "name": "Classic Pastrami",
+      "restaurant": "Otherside Deli",
+      "blurb": "yum yum eeatumup."
+   },
+   {
+      "id": 16,
+      "name": "Fried Bologna",
+      "restaurant": "The Silver Medal",
+      "blurb": "Best sandwich ever."
+   },
+   {
+      "id": 17,
+      "name": "BLT",
+      "restaurant": "Little Deli",
+      "blurb": "Best sandwich ever."
+   },
+   {
+      "id": 18,
+      "name": "BLT",
+      "restaurant": "Pho Vans",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 19,
+      "name": "Tuna Salad",
+      "restaurant": "Pho Vans",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 20,
+      "name": "Classic Pastrami",
+      "restaurant": "Pho Vans",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 21,
+      "name": "Harry's Perfect Pastrami",
+      "restaurant": "Pho Vans",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 22,
+      "name": "Classic Pastrami",
+      "restaurant": "Little Deli",
+      "blurb": "It slaps."
+   },
+   {
+      "id": 23,
+      "name": "Banh Mi",
+      "restaurant": "Pho Vans",
+      "blurb": "It's fucking good."
+   },
+   {
+      "id": 24,
+      "name": "Classic Pastrami",
+      "restaurant": "Otherside Deli",
+      "blurb": "Great for the summer."
+   },
+   {
+      "id": 25,
+      "name": "BLT",
+      "restaurant": "Little Deli",
+      "blurb": "yum yum eeatumup."
+   }
 ]
