+++ api/src/views/_nav.ejs
@@ -13,11 +13,11 @@
                     <li class="nav-item">
                         <a class="nav-link" href="/about">About</a>
                     </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#">Work</a>
+                    <li class="nav-item disabled">
+                        <a class="nav-link disabled">Work</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link disabled">Sandwiches</a>
+                        <a class="nav-link" href="/sandwiches">Sandwiches</a>
                     </li>
                 </ul>
             </div>
