+++ api/src/views/about.ejs
@@ -4,7 +4,7 @@
         <div class="container home-content">
           <h1 class="mt-3">Facts</h1>
           <p>
-            <span id="age">0</span> year old Austin based developer who wants to automate everything. Currently working on integrations for B2B and B2C systems for mondoshop.com.
+            <span id="age">0</span> year old Austin based developer who wants to automate everything. Currently working on integrations for B2B and DTC systems for mondoshop.com.
           </p>
           <p>
             I strongly believe in proper data hygeine and accesibility when it comes to working with as many endpoints as we do these days,
