+++ api/src/views/sandwiches.ejs
@@ -1,0 +1,38 @@
+<%- include('_header'); %>
+  <body>
+    <%- include('_nav'); %>
+        <div class="container home-content">
+          <h1 class="mt-3">Sandwiches</h1>
+          <p>
+          This is going to be the page for the random sandwich generator
+          </p>
+          <button type="button" class="btn btn-primary btn-lg">Random Sandwich</button>
+          <br>
+          <br>
+          <div>
+              <div class="card">
+                <div class="card-body">
+                  <h5 class="card-title">Sandwich Title</h5>
+                  <p class="card-text">
+                    Sandwich blurb.
+                  </p>
+                </div>
+                <ul class="list-group list-group-flush">
+                  <li class="list-group-item">Name</li>
+                  <li class="list-group-item">Restaurant</li>
+                </ul>
+                <div class="card-body">
+                  <a href="#" class="card-link disabled" onclick="getRandomSandwich()">Card link</a>
+                </div>
+              </div>
+            </div>
+
+        </div>
+    </div>
+    
+<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js" integrity="sha384-7+zCNj/IqJ95wo16oMtfsKbZ9ccEh31eOz1HGyDuCQ6wgnyJNSYdrPa03rtR1zdB" crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
+
+  </body>
+  <%- include('_footer'); %>
+</html>
